% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_sampleData.R
\name{aqs_annualData}
\alias{aqs_annualData}
\alias{aqs_annualData_bySite}
\alias{aqs_annualData_byCounty}
\alias{aqs_annualData_byState}
\alias{aqs_annualData_byCBSA}
\title{Extract annual average data data from the AQS API}
\usage{
aqs_annualData(aqs_user, endpoint = c("bySite", "byCounty", "byState",
  "byBox", "byCBSA"), param, bdate, edate, state = NULL, county = NULL,
  site = NULL, cbsa = NULL, minlat = NULL, maxlat = NULL,
  minlon = NULL, maxlon = NULL, ...)

aqs_annualData_bySite(aqs_user, param, bdate, edate, state, county, site,
  ...)

aqs_annualData_byCounty(aqs_user, param, bdate, edate, state, county, ...)

aqs_annualData_byState(aqs_user, param, bdate, edate, state, ...)

aqs_annualData_byCBSA(aqs_user, param, bdate, edate, cbsa, ...)
}
\arguments{
\item{aqs_user}{User information in the form of a list with \code{username} and \code{key}. See \code{\link{aqs_signup}}. If not provided, searches for an object of this name in the global environement.}

\item{endpoint}{Service endpoint}

\item{param}{Parameter code for which to extract data}

\item{bdate}{Begin date for data sample interval. Should be a string in format "YYYYMMDD"}

\item{edate}{End date for data sample interval. Should be a string in format "YYYYMMDD"}

\item{state}{Two-digit state code. Required when \code{endpoint} is one of \code{byState}, \code{byCounty}, \code{bySite}.}

\item{county}{Three-digit county code. Required when \code{endpoint} is \code{byCounty} or \code{bySite}.}

\item{site}{Three-digit site code. Required when \code{endpoint} is \code{bySite}.}

\item{cbsa}{Three-digit census-bureau statistical area (CBSA) code. Required when \code{endpoint} is \code{byCBSA}.}

\item{minlat}{Minimum latitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{maxlat}{Maximum latitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{minlon}{Minimum longitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{maxlon}{Maximum longitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{...}{Additional arguments passed to \code{\link{aqs_get}}}
}
\description{
Extract annual average data data from the AQS API
}
\seealso{
\code{\link{aqs_get}}

Other Data Query Functions: \code{\link{aqs_dailyData}},
  \code{\link{aqs_sampleData}}
}
\concept{Data Query Functions}
