% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_extract.R
\name{aqs_extract_data}
\alias{aqs_extract_data}
\title{Extract AQS data from JSON object}
\usage{
aqs_extract_data(r)
}
\arguments{
\item{r}{Response object}
}
\value{
A data frame
}
\description{
Extract AQS data from JSON object
}
\seealso{
\code{\link{aqs_get}}
}
