% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_generic.R
\name{aqs_get}
\alias{aqs_get}
\title{Get data from query to AQS API}
\usage{
aqs_get(service, endpoint, aqs_user, vars = NULL, run = TRUE,
  rawResponse = FALSE)
}
\arguments{
\item{service}{API service for which data are requested. See \code{\link{list_services}}}

\item{endpoint}{Service endpoint}

\item{aqs_user}{User information in the form of a list with \code{username} and \code{key}. See \code{\link{aqs_signup}}. If not provided, searches for an object of this name in the global environement.}

\item{vars}{List of variables for the endpoint request. See \code{\link{list_variables}}.}

\item{run}{Logical indicating whether request should be submitted (\code{TRUE}, default), or only the query URL returned (\code{FALSE}).}

\item{rawResponse}{Logical indicating that JSON response should be returned directly without checking or formatting. Used primarily in debugging.}
}
\description{
Get data from query to AQS API
}
\seealso{
\code{\link{aqs_sampleData}} \code{\link{aqs_list}}
}
