% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_list.R
\name{aqs_list}
\alias{aqs_list}
\alias{aqs_list_states}
\alias{aqs_list_counties}
\alias{aqs_list_sites}
\alias{aqs_list_parameters}
\title{List available services by endpoint}
\usage{
aqs_list(aqs_user, endpoint = c("states", "countiesByState",
  "sitesByCounty", "cbsas", "classes", "parametersByClass", "pqaos",
  "mas"), vars = NULL, ...)

aqs_list_states(aqs_user, ...)

aqs_list_counties(aqs_user, state, ...)

aqs_list_sites(aqs_user, state, county, ...)

aqs_list_parameters(aqs_user, pc, ...)
}
\arguments{
\item{aqs_user}{User information in the form of a list with \code{username} and \code{key}. See \code{\link{aqs_signup}}. If not provided, searches for an object of this name in the global environement.}

\item{endpoint}{Service endpoint}

\item{vars}{List of variables for the endpoint request. See \code{\link{list_variables}}.}

\item{...}{Additional arguments passed to \code{\link{aqs_get}}.}

\item{state}{Two-digit state code. Required when \code{endpoint} is one of \code{byState}, \code{byCounty}, \code{bySite}.}

\item{county}{Three-digit county code. Required when \code{endpoint} is \code{byCounty} or \code{bySite}.}

\item{pc}{Parameter class, for use when \code{endpoint} is \code{parametersByClass}}
}
\description{
List available services by endpoint
}
