% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_metadata.R
\name{aqs_metadata}
\alias{aqs_metadata}
\alias{aqs_isAvailable}
\alias{aqs_fieldsByService}
\title{Query metadata functions in AQS API}
\usage{
aqs_metadata(aqs_user, endpoint = c("isAvailable", "fieldsByService"),
  ...)

aqs_isAvailable(aqs_user)

aqs_fieldsByService(aqs_user, service = c("fieldsByService ", "list",
  "sampleData", "monitors", "sampleData", "dailyData", "annualData",
  "qaBlanks", "qaCollocated", "qaFlowRateVerifications",
  "qaFlowRateAudits", "qaOnePointQcRawData", "qaPepAudits"), ...)
}
\arguments{
\item{aqs_user}{User information in the form of a list with \code{username} and \code{key}. See \code{\link{aqs_signup}}. If not provided, searches for an object of this name in the global environement.}

\item{endpoint}{Service endpoint}

\item{...}{Additional arguments passed to \code{\link{aqs_get}}.}

\item{service}{API service for which data are requested. See \code{\link{list_services}}}
}
\description{
Query metadata functions in AQS API
}
\seealso{
\code{\link{aqs_get}} \code{\link{list_services}}
}
