% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_sampleData.R
\name{aqs_sampleData}
\alias{aqs_sampleData}
\alias{aqs_sampleData_bySite}
\alias{aqs_sampleData_byCounty}
\alias{aqs_sampleData_byState}
\alias{aqs_sampleData_byCBSA}
\title{Extract data from the AQS API}
\usage{
aqs_sampleData(
  aqs_user,
  endpoint = c("bySite", "byCounty", "byState", "byBox", "byCBSA"),
  param,
  bdate,
  edate,
  state = NULL,
  county = NULL,
  site = NULL,
  cbsa = NULL,
  minlat = NULL,
  maxlat = NULL,
  minlon = NULL,
  maxlon = NULL,
  ...
)

aqs_sampleData_bySite(aqs_user, param, bdate, edate, state, county, site, ...)

aqs_sampleData_byCounty(aqs_user, param, bdate, edate, state, county, ...)

aqs_sampleData_byState(aqs_user, param, bdate, edate, state, ...)

aqs_sampleData_byCBSA(aqs_user, param, bdate, edate, cbsa, ...)
}
\arguments{
\item{aqs_user}{User credentials. See \code{\link{aqs_signup}}}

\item{endpoint}{Endpoint for selecting data. Required variables vary by endpoint.}

\item{param}{Parameter code for which to extract data. Up to five codes may included in a single request.}

\item{bdate}{Begin date for data sample interval. Should be a string in format "YYYYMMDD"}

\item{edate}{End date for data sample interval. Should be a string in format "YYYYMMDD". Must be from the same year as \code{edate}.}

\item{state}{Two-digit state code. Required when \code{endpoint} is one of \code{byState}, \code{byCounty}, \code{bySite}.}

\item{county}{Three-digit county code. Required when \code{endpoint} is \code{byCounty} or \code{bySite}.}

\item{site}{Three-digit site code. Required when \code{endpoint} is \code{bySite}.}

\item{cbsa}{Three-digit census-bureau statistical area (CBSA) code. Required when \code{endpoint} is \code{byCBSA}.}

\item{minlat}{Minimum latitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{maxlat}{Maximum latitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{minlon}{Minimum longitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{maxlon}{Maximum longitude for the bounding box. Required when \code{endpoint} is \code{byBox}.}

\item{...}{Additional arguments passed to \code{\link{aqs_get}}}
}
\description{
Extract data from the AQS API
}
\details{
This queries the database using the "sampleData" endpoint. This will return any combination of measurements that meet the specified criteria. To query daily and annual summaries calculated by AQS, use \code{\link{aqs_dailyData}} and \code{\link{aqs_annualData}}.
}
\seealso{
\code{\link{aqs_get}}

Other Data Query Functions: 
\code{\link{aqs_annualData}()},
\code{\link{aqs_dailyData}()}
}
\concept{Data Query Functions}
