% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_signup.R
\name{aqs_signup}
\alias{aqs_signup}
\title{Sign up for AQS API}
\usage{
aqs_signup(email, ...)
}
\arguments{
\item{email}{Email address to register}

\item{...}{Additional arguments passed to \code{\link{aqs_get}}}
}
\description{
Submits a request to register email address and obtain key.
}
\details{
The AQS API requires an email address and key for all queries. The key is not used for authentication (as in a password), but it is used for identification.
The key value is emailed to the provided address, it is not return directly from the API.
}
