% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_vars.R
\name{list_services}
\alias{list_services}
\alias{list_endpoints}
\alias{list_required_vars}
\alias{list_optional_vars}
\title{List of Services and Endpoints for AQS API}
\usage{
list_services()

list_endpoints(service = NULL)

list_required_vars(endpoint = NULL)

list_optional_vars(endpoint = NULL)
}
\arguments{
\item{service}{Service for which endpoints should be returned. A value of NULL returns a list with all services.}

\item{endpoint}{Endpoint for which variables should be returned. A value of NULL returns a list with all services.}
}
\value{
A list containing the possible services or the possible endpoints for a specified service
}
\description{
List of Services and Endpoints for AQS API
}
\seealso{
aqs_list
}
